Pages & Header
Initialized pages and adjusted the RootLayout to better fit a sidebar and content. Improved folder structure.
