+++ app/analytics/page.tsx
@@ -1,6 +1,6 @@
 export default function Analytics() {
     return (
-        <div className="bg-[#242529] inline">
+        <div className="bg-[#242529]">
             Analytics
         </div>
     )
