+++ app/page.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image'
 
 export default function Home() {
   return (
-    <div className="bg-[#242529] inline">
+    <div className="bg-[#242529]">
       Home
     </div>
   )
