+++ app/reports/page.tsx
@@ -1,6 +1,6 @@
 export default function Reports() {
     return (
-        <div className="bg-[#242529] inline">
+        <div className="bg-[#242529]">
             Reports
         </div>
     )
