+++ app/settings/page.tsx
@@ -1,6 +1,6 @@
 export default function Settings() {
     return (
-        <div className="bg-[#242529] inline">
+        <div className="bg-[#242529]">
             Settings
         </div>
     )
