+++ components/navigation/sidebar.tsx
@@ -8,14 +8,14 @@ import {
     LogOut
 } from "lucide-react"
 
-import { NavBtn } from "@/components/nav-btns"
+import { NavBtn } from "@/components/navigation/nav-btns"
 import { usePathname } from "next/navigation"
 
 export const Sidebar = () => {
     const pathname = usePathname();
 
     return (
-        <aside className="hidden md:inline-flex flex-col gap-4 w-[368px] h-screen border-r border-primary-card p-[23px]">
+        <aside className="hidden md:inline-flex flex-col gap-4 w-[368px] h-screen border-r border-primary-card p-6">
             <header className="mb-4">
                 <h1 className="text-4xl">Dena Energy</h1>
             </header>
